Make the DAML trigger tests more generic
This is a first step towards making our tests more generic so we can
add more easily tests different triggers.
